I have organised our DTO files and repackaged the project for usability
