--- src/test/java/com/sparta/sc/DTO/SimWeatherDTO.java
@@ -1,178 +1,0 @@
-package com.sparta.sc.DTO;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.sparta.sc.utilities.SpeedConverter;
-
-import java.util.Arrays;
-import java.util.List;
-
-public class SimWeatherDTO {
-
-    private static final String[] weatherNames = new String[]{"Thunderstorm", "Drizzle", "Rain", "Snow", "Mist", "Smoke", "Haze", "Dust", "Fog", "Sand", "Dust", "Ash", "Squall", "Tornado", "Clear", "Clouds"};
-
-    @JsonProperty("visibility")
-    private int visibility;
-
-    @JsonProperty("timezone")
-    private int timezone;
-
-    @JsonProperty("main")
-    private Main main;
-
-    @JsonProperty("clouds")
-    private Clouds clouds;
-
-    @JsonProperty("sys")
-    private Sys sys;
-
-    @JsonProperty("dt")
-    private int dt;
-
-    @JsonProperty("coord")
-    private Coord coord;
-
-    @JsonProperty("weather")
-    private List<WeatherItem> weather;
-
-    @JsonProperty("name")
-    private String name;
-
-    @JsonProperty("cod")
-    private int cod;
-
-    @JsonProperty("id")
-    private int id;
-
-    @JsonProperty("base")
-    private String base;
-
-    @JsonProperty("wind")
-    private Wind wind;
-
-    public int getVisibility() {
-        return visibility;
-    }
-
-    public int getTimezone() {
-        return timezone;
-    }
-
-    public Main getMain() {
-        return main;
-    }
-
-    public Clouds getClouds() {
-        return clouds;
-    }
-
-    public Sys getSys() {
-        return sys;
-    }
-
-    public int getDt() {
-        return dt;
-    }
-
-    public Coord getCoord() {
-        return coord;
-    }
-
-    public List<WeatherItem> getWeather() {
-        return weather;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public int getCod() {
-        return cod;
-    }
-
-    public int getId() {
-        return id;
-    }
-
-    public String getBase() {
-        return base;
-    }
-
-    public Wind getWind() {
-        return wind;
-    }
-
-    // check if the lon is the same we passed as a parameter
-    public boolean isSameLon(double lon) {
-        return lon == getCoord().getLon();
-    }
-
-    // check if the lon is the same we passed as a parameter
-    public boolean isSameLat(double lat) {
-        return lat == getCoord().getLat();
-    }
-
-    public boolean isLonValid() {
-        return getCoord().getLon() >= -180 && getCoord().getLon() <= 180;
-    }
-
-    public boolean isLatValid() {
-        return getCoord().getLon() >= -90 && getCoord().getLon() <= 90;
-    }
-
-    public boolean isNameWeatherEmpty() {
-        return getWeather().get(0).getMain().equals("");
-    }
-
-    public boolean isNameWeatherValid() {
-        return Arrays.stream(weatherNames).anyMatch(name -> name.equals(getWeather().get(0).getMain()));
-    }
-
-    public boolean isDescriptionEmpty() {
-        return getWeather().get(0).getDescription().equals("");
-    }
-
-    public boolean isDegreeWindValid() {
-        return getWind().getDeg() >= 0 && getWind().getDeg() <= 360;
-    }
-
-    public double getWindInMps() {
-        return SpeedConverter.mphToMps(getWind().getSpeed());
-    }
-
-    public double getWindInMph() {
-        return SpeedConverter.mpsToMph(getWind().getSpeed());
-    }
-
-    public double getGustInMps() {
-        return SpeedConverter.mphToMps(getWind().getGust());
-    }
-
-    public double getGustInMph() {
-        return SpeedConverter.mpsToMph(getWind().getGust());
-    }
-
-    public boolean isCloudValid() {
-        return getClouds().getAll() >= 0 && getClouds().getAll() <= 100;
-    }
-
-    public boolean isNameEmpty() {
-        return getName().equals("");
-    }
-
-    public boolean isNameNull() {
-        return getName() == null;
-    }
-
-    public boolean isCountryTwoLetters() {
-        return getSys().getCountry().matches("[A-Za-z]{2}");
-    }
-
-    public boolean isStatusCode200() {
-        return getCod() == 200;
-    }
-
-    public boolean isIdBiggerThanO() {
-        return getId() > 0;
-    }
-
-}
